addes jwt-decode
