+++ frontend/package.json
@@ -12,6 +12,7 @@
   "dependencies": {
     "axios": "^1.7.7",
     "joi": "^17.13.3",
+    "jwt-decode": "^4.0.0",
     "react": "^18.3.1",
     "react-dom": "^18.3.1",
     "react-icons": "^5.3.0",
