+++ frontend/src/components/password/ResetPasswordForm.jsx
@@ -1,0 +1,7 @@
+
+
+const ResetPasswordForm = () => {
+  return <div>ResetPasswordForm</div>;
+};
+
+export default ResetPasswordForm;
