+++ frontend/src/pages/ChangePassword.jsx
@@ -1,0 +1,7 @@
+// import React from 'react'
+
+const ChangePassword = () => {
+  return <div>ChangePassword</div>;
+};
+
+export default ChangePassword;
