+++ frontend/src/pages/Home.jsx
@@ -2,9 +2,12 @@
 
 import axios from 'axios';
 import { useEffect, useState } from 'react';
+import { checkToken } from '../utils/checkToken';
+import { useNavigate } from 'react-router-dom';
 
 const Home = () => {
   const [blogs, setBlogs] = useState([]);
+  const navigate = useNavigate();
   const fetchBlog = async () => {
     try {
       const res = await axios.get('/api/blogs');
@@ -19,6 +22,12 @@ const Home = () => {
   useEffect(() => {
     fetchBlog();
   }, []);
+  // check token
+  useEffect(() => {
+    if (!checkToken()) {
+      navigate('/login');
+    }
+  }, [navigate]);
   return (
     <div>
       <h1>Home</h1>
