+++ frontend/src/pages/LandingPage.jsx
@@ -1,0 +1,7 @@
+// import React from 'react';
+
+const LandingPage = () => {
+  return <div>LandingPage</div>;
+};
+
+export default LandingPage;
