+++ frontend/src/pages/RegisterPage.jsx
@@ -1,0 +1,7 @@
+// import React from 'react';
+
+const RegisterPage = () => {
+  return <div>RegisterPage</div>;
+};
+
+export default RegisterPage;
