+++ frontend/src/pages/ResetPassword.jsx
@@ -1,0 +1,7 @@
+// import React from 'react';
+import '../styles/login.css';
+const ResetPassword = () => {
+  return <div>ResetPassword</div>;
+};
+
+export default ResetPassword;
